Category list Spinner moved to fragment with Quiz list. Some code clean up
