+++ app/src/main/java/com/example/quizer/userCabinet/User.java
@@ -7,35 +7,10 @@ public class User {
 
     private String password;
 
-    private int rating;
-
-    private String photoFileName;
-
-
-
-    public User() {
-    }
-
-    public User(String nickname,String password) {
+    public User(String nickname, String password) {
         this.nickname = nickname;
-        this.password= password;
-        this.rating=0;
-        this.photoFileName = "User_photo_"+nickname+".jpg";
-    }
-
-    public String getNickname() {
-        return nickname;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
         this.password = password;
     }
+
+
 }
