+++ app/src/main/res/layout/fragment_authorization.xml
@@ -74,7 +74,8 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:ems="10"
-                android:theme="@style/MyEditTextTheme" />
+                android:theme="@style/MyEditTextTheme"
+                android:inputType="text" />
 
         </TableRow>
