+++ app/src/main/res/values/colors.xml
@@ -7,5 +7,4 @@
     <color name="main_background">#ADC7CA</color>
 
     <color name="black">#000000</color>
-    <color name="edit_text_activated">#696666</color>
 </resources>
